Define ENET buffer using integer types with predefined width
The ENET IP is a 32 bit which expects to have the
buffers described using a specific format.
To avoid any problem caused by 32bit and 64 bit architecture
differences, the ENET buffer descriptors are defined using
integer types with predefined width.

+++ drivers/net/ethernet/freescale/fec.h
@@ -196,9 +196,9 @@
 #if defined(CONFIG_ARCH_MXC) || defined(CONFIG_SOC_IMX28) || \
     defined(CONFIG_ARCH_S32)
 struct bufdesc {
-	unsigned short cbd_datlen;	/* Data length */
-	unsigned short cbd_sc;	/* Control and status info */
-	unsigned long cbd_bufaddr;	/* Buffer address */
+	uint16_t cbd_datlen;	/* Data length */
+	uint16_t cbd_sc;	/* Control and status info */
+	uint32_t cbd_bufaddr;	/* Buffer address */
 };
 #else
 struct bufdesc {
@@ -210,11 +210,11 @@ struct bufdesc {
 
 struct bufdesc_ex {
 	struct bufdesc desc;
-	unsigned long cbd_esc;
-	unsigned long cbd_prot;
-	unsigned long cbd_bdu;
-	unsigned long ts;
-	unsigned short res0[4];
+	uint32_t cbd_esc;
+	uint32_t cbd_prot;
+	uint32_t cbd_bdu;
+	uint32_t ts;
+	uint16_t res0[4];
 };
 
 /*
